Attempt to fix commit 5e1269f
Visualize bScrollPos.

The goal is to fire networkDraw not only
when elements from both networkTransmitAccum
and networkOutputScroll are present, but
also when either one of them are present.

Previously, networkOutputScroll was:
accumB 0 ((+ 1) <$ filterJust eMouse)
which does not change value when eMouse is
Nothing.

The fix is to fire networkOutputScroll even
when eMouse is Nothing.

The idea is:
accumB 0 (maybe id (const (+ 1)) <$> eMouse)

However, it seems that accumB is returned
only when the accumulated value changes. So
that idea doesn't work, as networkDraw again
waits for elements from both
networkTransmitAccum and networkOutputScroll.

The attempt is to change the bScrollPos when
eMouse is both Nothing and Just, but by
different amounts.
accum 0 (maybe (+ 1) (const (+ 2)) <$> eMouse)

Future work:
Accumulate an anti-ScrollPos counter on
each Nothing eMouse in networkDraw, that
offsets the changes made to bScrollPos when
eMouse is Nothing.