Synchronize toy and real parsers and tests
Because real parser is very big and complicated,
keeping the toy parser around is good for quick
changes and fixes.

Toy parser is now a subset of real parser, so
that the outputs are the same. Now, you can do
`ghci toy/backend-toy.hs > out`, then compare
`out` with `test/parser/integer` and
`test/parser/bool` as a sanity check.

This makes it easier to ensure that subsequent
changes to the toy parser do not change its
output. Also, it is easier to compare the toy and
real implementations.